Update Metadata.swift
Fix the issue of Xcode 16 UI build failure

+++ Source/Metadata.swift
@@ -295,7 +295,7 @@ extension Metadata {
                 let cMangledTypeName = fieldRecords[i].mangledTypeName!
                 
                 let functionMap: [String: () -> Any.Type?] = [
-                    "function": { _getTypeByMangledNameInContext(cMangledTypeName, UInt(getMangledTypeNameSize(cMangledTypeName)), genericContext: self.contextDescriptorPointer, genericArguments: self.genericArgumentVector) }
+                    "function": { _getTypeByMangledNameInContext(cMangledTypeName, getMangledTypeNameSize(cMangledTypeName), genericContext: self.contextDescriptorPointer, genericArguments: self.genericArgumentVector) }
                 ]
                 if let function = functionMap["function"],let fieldType  = function() {
                     result.append(Property.Description(key: name, type: fieldType, offset: fieldOffsets[i]))
